Adds code copy and refactors formatting
