+++ src/example-helpers/component.js
@@ -27,11 +27,3 @@ export const component = instance => {
 	return Comp;
 };
 
-export const isEqual = (obj1, obj2) => {
-	if (obj1 === obj2) return true;
-	if (Object.keys(obj1).length !== Object.keys(obj2).length) return false;
-	for (let key in obj1) {
-		if (obj1[key] !== obj2[key]) return false;
-	}
-	return true;
-};
